♻️ Refactor user sign in logic into composables
